# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "ActionCreate": {
        "properties": {
          "auth": {
            "additionalProperties": true,
            "title": "Auth",
            "type": "object"
          },
          "host_allowlist": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Host Allowlist",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "spec": {
            "additionalProperties": true,
            "title": "Spec",
            "type": "object"
          }
        },
        "required": [
          "name",
          "spec"
        ],
        "title": "ActionCreate",
        "type": "object"
      },
      "ActionOut": {
        "properties": {
          "auth_type": {
            "title": "Auth Type",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "host_allowlist": {
            "items": {
              "type": "string"
            },
            "title": "Host Allowlist",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "operations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Operations",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "operations",
          "host_allowlist",
          "auth_type",
          "created_at"
        ],
        "title": "ActionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "McpServerCreate": {
        "properties": {
          "args": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "title": "Args",
            "type": "array"
          },
          "command": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "env": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Env",
            "type": "object"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "org": {
            "default": false,
            "title": "Org",
            "type": "boolean"
          },
          "transport": {
            "enum": [
              "stdio",
              "http"
            ],
            "title": "Transport",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "name",
          "transport"
        ],
        "title": "McpServerCreate",
        "type": "object"
      },
      "McpServerOut": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "has_secrets": {
            "title": "Has Secrets",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_status": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Status"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org": {
            "title": "Org",
            "type": "boolean"
          },
          "spec": {
            "additionalProperties": true,
            "title": "Spec",
            "type": "object"
          },
          "transport": {
            "title": "Transport",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "transport",
          "spec",
          "has_secrets",
          "enabled",
          "org",
          "last_status",
          "created_at"
        ],
        "title": "McpServerOut",
        "type": "object"
      },
      "McpServerPatch": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "McpServerPatch",
        "type": "object"
      },
      "McpToolOut": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "input_schema": {
            "additionalProperties": true,
            "title": "Input Schema",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "input_schema"
        ],
        "title": "McpToolOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Retinue",
    "version": "0.7.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/actions": {
      "get": {
        "operationId": "list_actions_api_actions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActionOut"
                  },
                  "title": "Response List Actions Api Actions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Actions",
        "tags": [
          "tools"
        ]
      },
      "post": {
        "operationId": "create_action_api_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Action",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/actions/{action_id}": {
      "delete": {
        "operationId": "delete_action_api_actions__action_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Action",
        "tags": [
          "tools"
        ]
      }
    },
...
    "/api/agents/{agent_id}/preflight": {
      "post": {
        "operationId": "preflight_api_agents__agent_id__preflight_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preflight Api Agents  Agent Id  Preflight Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preflight",
        "tags": [
          "tools"
        ]
      }
    },
...
    "/api/mcp/servers": {
      "get": {
        "operationId": "list_servers_api_mcp_servers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/McpServerOut"
                  },
                  "title": "Response List Servers Api Mcp Servers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Servers",
        "tags": [
          "tools"
        ]
      },
      "post": {
        "operationId": "create_server_api_mcp_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Server",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/mcp/servers/{server_id}": {
      "delete": {
        "operationId": "delete_server_api_mcp_servers__server_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Server",
        "tags": [
          "tools"
        ]
      },
      "patch": {
        "operationId": "patch_server_api_mcp_servers__server_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Server",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/mcp/servers/{server_id}/test": {
      "post": {
        "operationId": "test_server_api_mcp_servers__server_id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Test Server Api Mcp Servers  Server Id  Test Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Server",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/mcp/servers/{server_id}/tools": {
      "get": {
        "operationId": "server_tools_api_mcp_servers__server_id__tools_get",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Server Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/McpToolOut"
                  },
                  "title": "Response Server Tools Api Mcp Servers  Server Id  Tools Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Server Tools",
        "tags": [
          "tools"
        ]
      }
    },
...
  }
}